Refactored recommendation logic: moved buisness logic to model
