+++ models/Recommendation.js
@@ -1,15 +1,24 @@
+import { generateAIBookRecommendations } from '../services/openAiService.js';
+import { fetchBooks, getCovers } from '../services/googleBooksAPIWrapper.js';
 
-// what is a recommendation?
-// a book + reason + user ?
-//any other ideas?
+export async function createRecommendations(userPrompt) {
+  try {
+    const recommendations = await generateAIBookRecommendations(userPrompt);
+    if (!recommendations || !recommendations.books) {
+      return null;
+    }
+    // Fetch book covers from Google Books API
+    const titles = recommendations.books.map((book) => book.title);
+    const booksInfoFromGoogleBooks = await fetchBooks(titles);
+    const covers = getCovers(booksInfoFromGoogleBooks);
 
-// I agree, potentially its a list of books (4) with 
-// everything extra attached (reason and the cover) 
-// and the prompt that generated it ??
-
-//title
-//author
-//year
-//genre
-//reason
-//user
+    // Merge recommendations with cover images
+    return recommendations.books.map((book, index) => ({
+      recommendation: book,
+      cover: covers[index] || null,
+    }));
+  } catch (error) {
+    console.error('Error in Recommendation model: ', error);
+    throw new Error('Failed to create recommendations');
+  }
+}
